refactor: rename host input component
